feat: moving navbar location a new folder and change design visual component
